Remove old v0x0 API
