+++ app/controllers/api/v0/container_resource_quotas_controller.rb
@@ -1,7 +1,7 @@
 module Api
-  module V0x1
+  module V0
     class ContainerResourceQuotasController < ApplicationController
-      include Api::V0x1::Mixins::IndexMixin
+      include Api::V0::Mixins::IndexMixin
       include Api::V0::Mixins::ShowMixin
     end
   end
