+++ app/controllers/api/v0/mixins/index_mixin.rb
@@ -4,15 +4,16 @@ module Api
       module IndexMixin
         def index
           raise_unless_primary_instance_exists
-          render json: scoped(model.where(params_for_list))
+          render json: ManageIQ::API::Common::PaginatedResponse.new(
+            base_query: scoped(filtered.where(params_for_list)),
+            request: request,
+            limit: pagination_limit,
+            offset: pagination_offset
+          ).response
         end
 
         def scoped(relation)
-          if model.respond_to?(:taggable?) && model.taggable?
-            ref_schema = {model.tagging_relation_name => :tag}
-
-            relation = relation.includes(ref_schema).references(ref_schema)
-          elsif through_relation_klass
+          if through_relation_klass
             relation = relation.joins(through_relation_name)
           end
