+++ app/controllers/api/v0x1.rb
@@ -6,80 +6,26 @@ module Api
       end
     end
 
-    class AvailabilitiesController < Api::V0::AvailabilitiesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainersController < Api::V0x0::ContainersController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainerGroupsController < Api::V0x0::ContainerGroupsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainerImagesController < Api::V0x0::ContainerImagesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainerNodesController < Api::V0x0::ContainerNodesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainerProjectsController < Api::V0x0::ContainerProjectsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ContainerTemplatesController < Api::V0x0::ContainerTemplatesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class FlavorsController < Api::V0x0::FlavorsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class OrchestrationStacksController < Api::V0x0::OrchestrationStacksController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ServiceInstancesController < Api::V0x0::ServiceInstancesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ServiceOfferingIconsController < Api::V0x0::ServiceOfferingIconsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ServiceOfferingsController < Api::V0x0::ServiceOfferingsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class ServicePlansController < Api::V0x0::ServicePlansController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class SourcesController < Api::V0x0::SourcesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class TagsController < Api::V0x0::TagsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class VmsController < Api::V0x0::VmsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class VolumeAttachmentsController < Api::V0x0::VolumeAttachmentsController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class VolumeTypesController < Api::V0x0::VolumeTypesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
-
-    class VolumesController < Api::V0x0::VolumesController
-      include Api::V0x1::Mixins::IndexMixin
-    end
+    class AvailabilitiesController < Api::V0::AvailabilitiesController; end
+    class ContainersController < Api::V0::ContainersController; end
+    class ContainerGroupsController < Api::V0::ContainerGroupsController; end
+    class ContainerImagesController < Api::V0::ContainerImagesController; end
+    class ContainerNodesController < Api::V0::ContainerNodesController; end
+    class ContainerProjectsController < Api::V0::ContainerProjectsController; end
+    class ContainerResourceQuotasController < Api::V0::ContainerResourceQuotasController; end
+    class ContainerTemplatesController < Api::V0::ContainerTemplatesController; end
+    class FlavorsController < Api::V0::FlavorsController; end
+    class OrchestrationStacksController < Api::V0::OrchestrationStacksController; end
+    class ServiceInstancesController < Api::V0::ServiceInstancesController; end
+    class ServiceOfferingIconsController < Api::V0::ServiceOfferingIconsController; end
+    class ServiceOfferingsController < Api::V0::ServiceOfferingsController; end
+    class ServicePlansController < Api::V0::ServicePlansController; end
+    class SourcesController < Api::V0::SourcesController; end
+    class TagsController < Api::V0::TagsController; end
+    class TasksController < Api::V0::TasksController; end
+    class VmsController < Api::V0::VmsController; end
+    class VolumeAttachmentsController < Api::V0::VolumeAttachmentsController; end
+    class VolumeTypesController < Api::V0::VolumeTypesController; end
+    class VolumesController < Api::V0::VolumesController; end
   end
 end
