--- app/controllers/api/v0x1/tasks_controller.rb
@@ -1,19 +1,0 @@
-module Api
-  module V0x1
-    class TasksController < Api::V0x0::TasksController
-      include Api::V0x1::Mixins::IndexMixin
-
-      def update
-        model.update(params.require(:id), params_for_update)
-
-        messaging_client.publish_topic(
-          :service => "platform.topological-inventory.task-output-stream",
-          :event   => "Task.update",
-          :payload => params_for_update.to_h.merge("task_id" => params.require(:id))
-        )
-
-        head :no_content
-      end
-    end
-  end
-end
