+++ spec/controllers/api/v0x1/container_projects_controller_spec.rb
@@ -1,4 +1,4 @@
 RSpec.describe Api::V0x1::ContainerProjectsController, :type => :request do
-  it("Uses IndexMixin") { expect(described_class.instance_method(:index).owner).to eq(Api::V0x1::Mixins::IndexMixin) }
+  it("Uses IndexMixin") { expect(described_class.instance_method(:index).owner).to eq(Api::V0::Mixins::IndexMixin) }
   it("Uses ShowMixin")  { expect(described_class.instance_method(:show).owner).to eq(Api::V0::Mixins::ShowMixin) }
 end
