+++ spec/controllers/api/v0x1/service_plans_controller_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Api::V0x1::ServicePlansController, :type => :request do
   include ::Spec::Support::TenantIdentity
   let(:headers) { {"x-rh-identity" => identity} }
 
-  it("Uses IndexMixin") { expect(described_class.instance_method(:index).owner).to eq(Api::V0x1::Mixins::IndexMixin) }
+  it("Uses IndexMixin") { expect(described_class.instance_method(:index).owner).to eq(Api::V0::Mixins::IndexMixin) }
   it("Uses ShowMixin")  { expect(described_class.instance_method(:show).owner).to eq(Api::V0::Mixins::ShowMixin) }
 
   describe "#order" do
