+++ spec/controllers/api/v0x1/tasks_controller_spec.rb
@@ -1,7 +1,7 @@
 require "manageiq-messaging"
 
 RSpec.describe Api::V0x1::TasksController, :type => :request do
-  it("Uses IndexMixin")   { expect(described_class.instance_method(:index).owner).to eq(Api::V0x1::Mixins::IndexMixin) }
+  it("Uses IndexMixin")   { expect(described_class.instance_method(:index).owner).to eq(Api::V0::Mixins::IndexMixin) }
   it("Uses ShowMixin")    { expect(described_class.instance_method(:show).owner).to eq(Api::V0::Mixins::ShowMixin) }
 
   include ::Spec::Support::TenantIdentity
